refactor : presenterTest를 viewModel 테스트로 변경
